+++ contao/config/config.php
@@ -7,6 +7,6 @@ array_insert($GLOBALS['BE_MOD']['content'], 1, [
     ]
 ]);
 
-$GLOBALS['FE_MOD']['application']['articlefilter']            = '\ContaoArticleFilter\ModuleArticleFilter';
-$GLOBALS['FE_MOD']['application']['articlefilter_links']      = '\ContaoArticleFilter\ModuleFilterLinks';
-$GLOBALS['FE_MOD']['application']['articlefilter_results']    = '\ContaoArticleFilter\ModuleFilterResults';
+$GLOBALS['FE_MOD']['application']['articlefilter']            = '\ContaoArticleFilter\Modules\ModuleArticleFilter';
+$GLOBALS['FE_MOD']['application']['articlefilter_links']      = '\ContaoArticleFilter\Modules\ModuleFilterLinks';
+$GLOBALS['FE_MOD']['application']['articlefilter_results']    = '\ContaoArticleFilter\Modules\ModuleFilterResults';
